requirements.txt
Django
djangorestframework
psycopg2

flake8>=3.6.0,<3.7.0
